Port to MacOS X (Fixes #93)
Installer script creates self-contained Trelby.app package,
which should be dragged to /Applications.  To run directly
from source without packaging, you will need to install lxml,
Python 2.7, and wxPython 2.9 to your Mac.  Then just symlink
/Applications/Trelby.app to the source directory and it should run.

The Makefile command "make dist-osx" uses py2app, which must be
installed if you want to create a self-contained package.  Once the
package is built, the result can be found in /dist.  Trelby.app
can be dragged to /Applications and should run even on Macs with older
Python or without lxml or wxPython previously installed.

If you want to install to /Applications automatically, just use:

make install-osx

This will also create a folder on the desktop with the license,
sample documents, and some other helpful files, including a simlink
to the app in the /Applications folder.

Thanks to anilgulecha and Max F for getting this port started at
http://tinyurl.com/8z8nmbb

Features:

*  Runs on MacOS X (tested on 10.6.x/10.8.x)
*  Runs as native app at /Applications/Trelby.app
*  Self-contained- no additional libraries/python required
*  Moved menu items to better match Mac conventions
*  Recognizes native Quit instruction
*  Modify "ALT"->"CMD" in shortcut settings
*  Launches Viewer.app for print feature
*  Makes the manual xslt xml->html conversion compatible with OS X
*  Adds MacOS X native warnings for missing/old Python/wx (for devs)
*  MacOS native warning to run in /Applications (for users)
*  Sets Desktop as default script destination (also fixes for unix)
*  Uses MacOS X native color palette
*  Adds installation scripts to generate standalone Trelby.app + archive
*  Adds MacOS X build instructions
*  Uses 256x256 Trelby icon as app & document icons
*  Auto-opens Trelby document on double-click from Finder

This is a first pass at this port, so there may be bugs and things
to work out.