Make paper size default to "Letter" when run in Americas
Determine if the local timezone is between GMT-10 and GMT-3,
or roughly from HAST (HI) in the west to NFT (T-N) in the east.
This range includes the US, Canada, Mexico, and much of Latin
America.  If the timezone falls in that range, use "Letter"
default paper size rather than "A4" ISO standard, as this is
a common standard in most (but not all) of these countries.

There are many, many page standards in use, and at some
point it may be a good idea to match specific country codes
to page sizes, but this is a decent rough method to select
a default, especially for newer North American industry users
who may be confused by the default A4 paper size and not
immediately know what's wrong or how to fix it via the
settings.