Associate Trelby file type for Linux (Fixes #265)
Creates "application/trelby" MIME type on *nix-based
systems following freedesktop.org standards.  This
results in documents ending in ".trelby" having a
custom icon and auto-opening w/double-click.

Commit summary:

* Sets up "application/trelby" MIME type
* Any file ending in ".trelby" is associated
* Sets up five icon sizes
* Automate install by generated .deb packages
* Auto-check for xdg-* support
* If Trelby is uninstalled, filetype association
  should be automatically removed

Note that this is installed for Debian-type installations.
Just use a simple "make deb" and then dpkg -i the .deb file.

Support can be easily added to support local "make install"
This was not implemented here as the Makefile is still
installing to /opt and it is unclear if this is the final
directory or if there are plans to use debian-style
locations.  In any event, the Makefile to support /opt/trelby
is already written and can be added in another commit.

This is a first pass, but it seems to work fine.  However,
there always may be issues that need addressing.